+++ src/sovereign_agent/database.py
@@ -15,6 +15,7 @@ from sovereign_agent.assistant_schema import SCHEMA_21 as MIGRATION_21
 from sovereign_agent.assistant_schema import SCHEMA_22 as MIGRATION_22
 from sovereign_agent.assistant_schema import SCHEMA_23 as MIGRATION_23
 from sovereign_agent.assistant_schema import SCHEMA_24 as MIGRATION_24
+from sovereign_agent.assistant_schema import SCHEMA_25 as MIGRATION_25
 
 MIGRATION_1 = """
 PRAGMA foreign_keys = ON;
@@ -828,6 +829,7 @@ MIGRATIONS: tuple[tuple[int, str], ...] = (
     (22, MIGRATION_22),
     (23, MIGRATION_23),
     (24, MIGRATION_24),
+    (25, MIGRATION_25),
 )